The new coincident/server
