+++ src/server/constants.js
@@ -1,0 +1,4 @@
+import { MAIN, WORKER } from '../window/constants.js';
+
+export const MAIN_WS = MAIN + '-ws';
+export const WORKER_WS = WORKER + '-ws';
