+++ src/server/worker.js
@@ -1,0 +1,5 @@
+import coincident from '../window/worker.js';
+
+export default async options => {
+  // TODO
+};
